Updated pop_remap to allow numerics instead of just floats. Removed an extra include file from dtrendW.c.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6135 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/wrapper.c
@@ -1613,11 +1613,11 @@ void NclAddUserFuncs(void)
     nargs = 0;
     args = NewArgs(5);
 
-    SetArgTemplate(args,nargs,"float",1,NclANY);nargs++;
-    SetArgTemplate(args,nargs,"float",2,NclANY);nargs++;
+    SetArgTemplate(args,nargs,"numeric",1,NclANY);nargs++;
+    SetArgTemplate(args,nargs,"numeric",2,NclANY);nargs++;
     SetArgTemplate(args,nargs,"integer",1,NclANY);nargs++;
     SetArgTemplate(args,nargs,"integer",1,NclANY);nargs++;
-    SetArgTemplate(args,nargs,"float",1,NclANY);nargs++;
+    SetArgTemplate(args,nargs,"numeric",1,NclANY);nargs++;
 
     NclRegisterProc(pop_remap_W,args,"pop_remap",nargs);
